Migrate updateMTOReviewedBillableWeightsAt
- extract out the event and related test
- uncomment updateMTOReviewedBillableWeightsAt in index.js
- remove the extracted event and test from moveHistoryEventTemplate and the test file
